Simplify other tests by using all the same dict
