+++ tests/unit_tests/test_simulator.py
@@ -6,7 +6,6 @@ import shutil
 from copy import copy
 from pathlib import Path
 
-import astropy.units as u
 import pytest
 
 import simtools.utils.general as gen
@@ -26,46 +25,8 @@ def label():
 
 
 @pytest.fixture
-def simulator_config_data(tmp_test_directory):
-    return {
-        "common": {
-            "site": "North",
-            "layout_name": "test-layout",
-            "data_directory": f"{str(tmp_test_directory)}/test-output",
-            "zenith": 20 * u.deg,
-            "azimuth": 0 * u.deg,
-            "primary": "gamma",
-        },
-        "showers": {
-            "eslope": -2.5,
-            "viewcone": [0 * u.deg, 0 * u.deg],
-            "nshow": 10,
-            "erange": [100 * u.GeV, 1 * u.TeV],
-            "cscat": [10, 1400 * u.m, 0],
-            "run_list": [3, 4],
-            "run_range": [6, 10],
-        },
-        "array": {
-            "model_version": "Prod5",
-            "default": {"LST": "D234", "MST": "NectarCam-D"},
-            "LST-01": "1",
-        },
-    }
-
-
-@pytest.fixture
-def array_config_data(simulator_config_data):
-    return simulator_config_data["common"] | simulator_config_data["array"]
-
-
-@pytest.fixture
 def input_file_list():
     return ["run1", "abc_run22", "def_run02_and"]
-
-
-@pytest.fixture
-def shower_config_data(simulator_config_data):
-    return simulator_config_data["common"] | simulator_config_data["showers"]
 
 
 @pytest.fixture
